+++ lib/easy-rest/handlers/error-request-handler.d.ts
@@ -1,4 +1,0 @@
-import { Response, Request } from "express";
-export interface IErrorRequestHandler {
-    (err: any, req: Request, res: Response): Promise<any>;
-}