+++ lib/index.d.ts
@@ -2,7 +2,6 @@ export * from './easy-rest/api';
 export * from './easy-rest/caching';
 export * from './easy-rest/core';
 export * from './easy-rest/decorators';
-export * from './easy-rest/handlers';
 export * from './easy-rest/metadata';
 export * from './easy-rest/security';
 export * from './easy-rest/util';