+++ src/easy-rest/core/module-loader.ts
@@ -1,6 +1,8 @@
 import * as Glob from 'glob';
 import {IControllerConstructor} from "../api/api-controller";
 
+let loaded: string[] = [];
+
 //TODO: don't use static
 export class ModuleLoader {
   static controllers: IControllerConstructor[] = [];
@@ -9,7 +11,14 @@ export class ModuleLoader {
     let files = Glob.sync(pattern);
 
     for (let i = 0; i < files.length; i++) {
-      require(files[i]);
+      let file = files[i];
+
+      if (loaded.indexOf(file) !== -1) {
+        continue;
+      }
+
+      require(file);
+      loaded.push(file);
     }
   }
 }