+++ src/easy-rest/handlers/request-handler.ts
@@ -1,9 +1,0 @@
-import {Response, Request} from "express";
-
-export interface IRequestHandler {
-  /**
-   * Resolve to continue with next handler
-   * Reject to continue with error handlers
-   */
-  (req: Request, res: Response): Promise<void>;
-}